% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abs_tools.R
\name{abs_wide_to_long}
\alias{abs_wide_to_long}
\title{Convert a wide ABS table to a long one, adding in a base set of variables as appropriate}
\usage{
abs_wide_to_long(abs_table)
}
\arguments{
\item{abs_table}{An abs data.frame}
}
\value{
Returns a data.frame in long format with augmented columns.
}
\description{
Convert a wide ABS table to a long one, adding in a base set of variables as appropriate
}
\details{
Convert a wide ABS table to a long one, adding in a base set of variables as appropriate
}
\examples{
abs_table <- read_abs('BCP', 'B46', 'SA4')
abs_wide_to_long(abs_table)
}

