% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asgs.info.R
\docType{data}
\name{asgs.info}
\alias{asgs.info}
\title{ABS Australian Census 2011: ASGS code info}
\format{List indexed by level name
\describe{
 \item{volume}{ABS volume number e.g. 1270055001}
 \item{columns}{Column names in the shapefile}
}}
\source{
Based on Australian Bureau of Statistics data \url{http://www.abs.gov.au/}
}
\usage{
asgs.info
}
\description{
ABS Australian Census 2011: ASGS code info
}
\details{
ASGS code info for the 2011 Australian Census Data.
}
\keyword{datasets}

