% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asgs.mb.alloc.R
\docType{data}
\name{asgs.mb.alloc}
\alias{asgs.mb.alloc}
\title{ABS Australian Census 2011: ASGS area hierarchy. Immediate enclosing regions of MeshBlocks i.e. SA1 and LGA}
\format{List indexed by level name
\describe{
 \item{volume}{ABS volume number e.g. 1270055001}
 \item{columns}{Column names in the shapefile}
}}
\source{
Based on Australian Bureau of Statistics data \url{http://www.abs.gov.au/}
}
\usage{
asgs.mb.alloc
}
\description{
ABS Australian Census 2011: ASGS area hierarchy. Immediate enclosing regions of MeshBlocks i.e. SA1 and LGA
}
\details{
ABS Australian Census 2011: ASGS area hierarchy. Immediate enclosing regions of MeshBlocks i.e. SA1 and LGA
}
\keyword{datasets}

