% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asgs.sa1.alloc.R
\docType{data}
\name{asgs.sa1.alloc}
\alias{asgs.sa1.alloc}
\title{ABS Australian Census 2011: ASGS area hierarchy. Enclosing regions of SA1}
\format{List indexed by level name
\describe{
 \item{volume}{ABS volume number e.g. 1270055001}
 \item{columns}{Column names in the shapefile}
}}
\source{
Based on Australian Bureau of Statistics data \url{http://www.abs.gov.au/}
}
\usage{
asgs.sa1.alloc
}
\description{
ABS Australian Census 2011: ASGS area hierarchy. Enclosing regions of SA1
}
\details{
ABS Australian Census 2011: ASGS area hierarchy. Enclosing regions of SA1
}
\keyword{datasets}

