% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asgs.sa2.R
\docType{data}
\name{asgs.sa2}
\alias{asgs.sa2}
\title{ABS Australian Census 2011: ASGS area hierarchy. Immediate enclosing regions of SA2}
\format{List indexed by level name
\describe{
 \item{volume}{ABS volume number e.g. 1270055001}
 \item{columns}{Column names in the shapefile}
}}
\source{
Based on Australian Bureau of Statistics data \url{http://www.abs.gov.au/}
}
\usage{
asgs.sa2
}
\description{
ABS Australian Census 2011: ASGS area hierarchy. Immediate enclosing regions of SA2
}
\details{
ABS Australian Census 2011: ASGS area hierarchy. Immediate enclosing regions of SA2
}
\keyword{datasets}

