% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asgs_tools.R
\name{asgs_extract_named_region}
\alias{asgs_extract_named_region}
\title{Extract a named region from an ASGS shapefile}
\usage{
asgs_extract_named_region(region_name, boundaries)
}
\arguments{
\item{region_name}{character string of the name to extract}

\item{boundaries}{ASGS shapefile (SpatialPolygonsDataFrame)}
}
\description{
Extract a named region from an ASGS shapefile
}
\details{
Extract a named region from an ASGS shapefile
}

