% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asgs_tools.R
\name{asgs_get_all_spcentroids}
\alias{asgs_get_all_spcentroids}
\title{Calculate a SpatialPointsDataFrame of the centroids of every region in an ASGS shapefile}
\usage{
asgs_get_all_spcentroids(boundaries)
}
\arguments{
\item{boundaries}{ASGS shapefile (SpatialPolygonsDataFrame)}
}
\value{
SpatialPointsDataFrame contains the centroids of all the regions in the given 'boundaries'
}
\description{
Calculate a SpatialPointsDataFrame of the centroids of every region in an ASGS shapefile
}
\details{
Calculate a SpatialPointsDataFrame of the centroids of every region in an ASGS shapefile
}

