% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asgs_tools.R
\name{asgs_get_spcentroid_by_name}
\alias{asgs_get_spcentroid_by_name}
\title{Return the centroid of a named region of a shapefile as SpatialPoints}
\usage{
asgs_get_spcentroid_by_name(name, boundaries)
}
\arguments{
\item{name}{character string of region to extract}

\item{boundaries}{ASGS shapefile (SpatialPolygonsDataFrame)}
}
\value{
SpatialPoints structure containing (longitude, latitude) of region centroid
}
\description{
Return the centroid of a named region of a shapefile as SpatialPoints
}
\details{
Return the centroid of a named region of a shapefile as SpatialPoints
}

