% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asgs_tools.R
\name{asgs_load_shapefile}
\alias{asgs_load_shapefile}
\title{Load an ASGS shapefile}
\usage{
asgs_load_shapefile(level, shapefiles_dir = paste0(Sys.getenv("HOME"),
  "/projectsdata/ABS2011/ESRI-Shapefile/"))
}
\arguments{
\item{level}{Structure level e.g. GCCSA, SSC, POA etc}

\item{shapefiles_dir}{location of shapefile directories}
}
\value{
ASGS shapefile (SpatialPolygonsDataFrame)
}
\description{
Load an ASGS shapefile supplied by the ABS
}
\details{
Load an ASGS shapefile
}

