% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asgs_tools.R
\name{asgs_select_bbox}
\alias{asgs_select_bbox}
\title{Select only the geometry within the given bounding box}
\usage{
asgs_select_bbox(boundaries, min_long, min_lat, max_long, max_lat)
}
\arguments{
\item{boundaries}{ASGS shapefile (SpatialPolygonsDataFrame)}

\item{min_long}{min longitude}

\item{min_lat}{min latitude}

\item{max_long}{max longitude}

\item{max_lat}{max latitude}
}
\value{
SpatialPolygonsDataFrame of bounded region
}
\description{
Select only the geometry within the given bounding box
}
\details{
Select only the geometry within the given bounding box
}

