% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asgs_tools.R
\name{asgs_simplify_boundaries}
\alias{asgs_simplify_boundaries}
\title{Simplify ASGS data}
\usage{
asgs_simplify_boundaries(boundaries, simplification_tolerance = 0.01,
  preserve_topology = TRUE)
}
\arguments{
\item{boundaries}{ASGS shapefile (SpatialPolygonsDataFrame)}

\item{simplification_tolerance}{higher numbers mean more simplification}

\item{preserve_topology}{preserve topology}
}
\value{
SpatialPolygonsDataFrame of simplified region
}
\description{
Simplify ASGS data
}
\details{
Simplify ASGS data
}

