% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geog.desc.R
\docType{data}
\name{geog.desc}
\alias{geog.desc}
\title{ABS Australian Census 2011: Geographical Descriptors}
\format{Data frame with columns
\describe{
 \item{level}{AUS, SSC, STE etc}
 \item{region_id}{ID for region within a statistical area}
 \item{label}{Full text label}
 \item{area}{Area of region in square kms}
}}
\source{
Based on Australian Bureau of Statistics data \url{http://www.abs.gov.au/websitedbs/censushome.nsf/home/datapacksdetails/$file/2011Census_geog_desc_1st_and_2nd_release.csv}
}
\usage{
geog.desc
}
\description{
ABS Australian Census 2011 Geographical Descriptors
}
\details{
Metadata concerning the description of the different statistical regions for the 2011 Australian Census Data.
}
\keyword{datasets}

