% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asgs_tools.R
\name{merge_asgs_abs}
\alias{merge_asgs_abs}
\title{Merge the boundaries data and the absdata into a data.frame ready for ggplot}
\usage{
merge_asgs_abs(boundaries, absdata)
}
\arguments{
\item{boundaries}{ASGS shapefile (SpatialPolygonsDataFrame)}

\item{absdata}{data.frame of ABS data in long format}
}
\value{
data.frame ready for ggplot2 plotting
}
\description{
Merge the boundaries data and the absdata into a data.frame ready for ggplot
By doing an inner_join here, we only return the plotting data that is the
region at the intersection of boundaries and absdata
}

