% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{read_abs}
\alias{read_abs}
\title{Load an ABS DataPack file}
\usage{
read_abs(profile, table, level)
}
\arguments{
\item{profile}{The short code for the profile. e.g. 'BCP' for Basic Community Profile.}

\item{table}{Table number excluding suffix. e.g. 'B01', 'B08'}

\item{level}{The level of the statistcal area: [AUS, CED, GCCSA, IARE, ILOC, IREG, LGA, POA, RA, SA1, SA2, SA3, SA4, SED, SLA, SOS, SOSR, SSC, STE, SUA, UCL]}
}
\value{
Returns the full path to the CSV file containing the nominated data.
}
\description{
Given a profile, table & level, read the relevant ABS data into a data.frame
}
\details{
Because of the size of the data, many tables are split over multiple files and numbered
BxxA, BxxB, BxxC etc.  Once a DataPack has over 200 columns (excluding the region_id column),
further columns are spilled into the next file.
\code{read_abs} will automatically read these multiple files and
merge the data into one really really wide data.frame
}
\examples{
read_abs('BCP', 'B46', 'AUS')
}

