% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split.variables.R
\docType{data}
\name{split.variables}
\alias{split.variables}
\title{ABS Australian Census 2011: Basic splitting of column names into variables for each of the BCP tables}
\format{List of data.frames indexed by table name e.g. 'B01'
\describe{
 \item{colname}{The original (long) column name from the ABS CSV file}
 \item{...}{1 or more columns containing the parsed/split variables from this column name}
}}
\source{
Based on Australian Bureau of Statistics data \url{http://www.abs.gov.au/}
}
\usage{
split.variables
}
\description{
ABS Australian Census 2011: Basic splitting of column names into variables for each of the BCP tables
}
\details{
A basic splitting of the column names for each table into a more useful set of variables.  e.g. "Female_20-24_12" might become: gender='Female', age='20-24', cars_owned=12.

See the vignettes for more examples on usage.
}
\keyword{datasets}

