% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tableconfig.R
\docType{data}
\name{tableconfig}
\alias{tableconfig}
\title{ABS Australian Census 2011: Some metainfo for BCP table}
\format{List of metainformation indexed by table name e.g. 'B01'
\describe{
 \item{table}{Table Name e.g. 'B01'}
 \item{desc}{Short description of table contents}
 \item{patterns}{Regular expressions used for splitting column names into variables. Don't look at these unless you want to be shocked.}
 \item{stats}{The new variables to be created by splitting up the column names into components}
}}
\source{
Based on Australian Bureau of Statistics data \url{http://www.abs.gov.au/}
}
\usage{
tableconfig
}
\description{
ABS Australian Census 2011 Table Configuration
}
\details{
Metadata concerning the description of the BCP data tables for the 2011 Australian Census Data.  Note that this is mostly for internal use in order to orchestrate the parsing of the raw tables into R datasets
}
\keyword{datasets}

